#65 introduce generic extendable extractor
